more reorganization, move visualization tools into boxatlas.py
